requirements.txt
IMAPClient==2.1.0
python-dotenv==0.10.3
six==1.12.0
